Refactor portfolio with default cursor and smooth scroll
