+++ package.json
@@ -15,6 +15,7 @@
     "@react-three/fiber": "^8.17.10",
     "framer-motion": "^11.13.5",
     "framer-motion-3d": "^11.2.0",
+    "lenis": "^1.0.38",
     "lucide-react": "^0.468.0",
     "next": "14.2.20",
     "react": "^18",
